+++ src/voidcode/provider/__init__.py
@@ -1,0 +1,66 @@
+from __future__ import annotations
+
+from .config import (
+    ProviderFallbackConfig,
+    parse_provider_fallback_payload,
+    serialize_provider_fallback_config,
+)
+from .errors import (
+    SingleAgentContextLimitError,
+    SingleAgentProviderError,
+    classify_provider_error,
+    format_fallback_exhausted_error,
+    format_invalid_provider_config_error,
+)
+from .models import (
+    ProviderModelSelection,
+    ResolvedProviderChain,
+    ResolvedProviderConfig,
+    ResolvedProviderModel,
+)
+from .protocol import (
+    ModelProvider,
+    ProviderExecutionError,
+    SingleAgentProvider,
+    SingleAgentTurnRequest,
+    SingleAgentTurnResult,
+    StubSingleAgentProvider,
+)
+from .registry import ModelProviderRegistry, StaticModelProvider
+from .resolution import (
+    resolve_provider_chain,
+    resolve_provider_config,
+    resolve_provider_model,
+)
+from .snapshot import (
+    parse_resolved_provider_snapshot,
+    resolved_provider_snapshot,
+)
+
+__all__ = [
+    "ModelProvider",
+    "ModelProviderRegistry",
+    "ProviderExecutionError",
+    "ProviderFallbackConfig",
+    "ProviderModelSelection",
+    "ResolvedProviderChain",
+    "ResolvedProviderConfig",
+    "ResolvedProviderModel",
+    "SingleAgentProvider",
+    "SingleAgentTurnRequest",
+    "SingleAgentTurnResult",
+    "SingleAgentContextLimitError",
+    "SingleAgentProviderError",
+    "StaticModelProvider",
+    "StubSingleAgentProvider",
+    "classify_provider_error",
+    "format_fallback_exhausted_error",
+    "format_invalid_provider_config_error",
+    "parse_resolved_provider_snapshot",
+    "parse_provider_fallback_payload",
+    "resolve_provider_chain",
+    "resolve_provider_config",
+    "resolve_provider_model",
+    "resolved_provider_snapshot",
+    "serialize_provider_fallback_config",
+]
