+++ src/voidcode/provider/config.py
@@ -1,0 +1,86 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import cast
+
+
+@dataclass(frozen=True, slots=True)
+class ProviderFallbackConfig:
+    preferred_model: str
+    fallback_models: tuple[str, ...] = ()
+
+
+def parse_provider_fallback_payload(
+    raw_provider_fallback: object,
+    *,
+    source: str,
+) -> ProviderFallbackConfig | None:
+    if raw_provider_fallback is None:
+        return None
+    if not isinstance(raw_provider_fallback, dict):
+        raise ValueError(f"{source} must be an object when provided")
+
+    payload = cast(dict[str, object], raw_provider_fallback)
+    preferred_model = payload.get("preferred_model")
+    if not isinstance(preferred_model, str):
+        raise ValueError(f"{_nested_config_field(source, 'preferred_model')} must be a string")
+    fallback_models = _parse_string_list(
+        payload.get("fallback_models"),
+        field_path=_nested_config_field(source, "fallback_models"),
+    )
+    ordered_models = (preferred_model, *fallback_models)
+    if len(set(ordered_models)) != len(ordered_models):
+        raise ValueError("provider fallback chain must not contain duplicate models")
+    return ProviderFallbackConfig(
+        preferred_model=preferred_model,
+        fallback_models=fallback_models,
+    )
+
+
+def serialize_provider_fallback_config(
+    provider_fallback: ProviderFallbackConfig | None,
+) -> dict[str, object] | None:
+    if provider_fallback is None:
+        return None
+    return {
+        "preferred_model": provider_fallback.preferred_model,
+        "fallback_models": list(provider_fallback.fallback_models),
+    }
+
+
+def _parse_string_list(raw_value: object, *, field_path: str) -> tuple[str, ...]:
+    if raw_value is None:
+        return ()
+    if not isinstance(raw_value, list):
+        raise ValueError(
+            f"{_format_runtime_config_field_error(field_path)} must be an array when provided"
+        )
+
+    raw_items = cast(list[object], raw_value)
+    parsed_items: list[str] = []
+    for index, item in enumerate(raw_items):
+        if not isinstance(item, str):
+            raise ValueError(
+                f"{_format_runtime_config_field_error(f'{field_path}[{index}]')} must be a string"
+            )
+        parsed_items.append(item)
+    return tuple(parsed_items)
+
+
+def _nested_config_field(source: str, nested: str) -> str:
+    runtime_field_prefix = "runtime config field '"
+    if source.startswith(runtime_field_prefix) and source.endswith("'"):
+        base_field = source[len(runtime_field_prefix) : -1]
+        return f"runtime config field '{base_field}.{nested}'"
+    return f"{source}.{nested}"
+
+
+def _format_runtime_config_field_error(field_path: str) -> str:
+    runtime_field_prefix = "runtime config field '"
+    if field_path.startswith(runtime_field_prefix):
+        if field_path.endswith("'"):
+            return field_path
+        if "'[" in field_path:
+            base, suffix = field_path[len(runtime_field_prefix) :].split("'[", maxsplit=1)
+            return f"{runtime_field_prefix}{base}[{suffix}'"
+    return f"runtime config field '{field_path}'"
