+++ src/voidcode/provider/errors.py
