+++ src/voidcode/provider/models.py
@@ -1,0 +1,39 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from .config import ProviderFallbackConfig
+from .protocol import ModelProvider
+
+
+@dataclass(frozen=True, slots=True)
+class ProviderModelSelection:
+    raw_model: str | None = None
+    provider: str | None = None
+    model: str | None = None
+
+
+@dataclass(frozen=True, slots=True)
+class ResolvedProviderModel:
+    selection: ProviderModelSelection = ProviderModelSelection()
+    provider: ModelProvider | None = None
+
+
+@dataclass(frozen=True, slots=True)
+class ResolvedProviderChain:
+    preferred: ResolvedProviderModel = ResolvedProviderModel()
+    fallbacks: tuple[ResolvedProviderModel, ...] = ()
+    all_targets: tuple[ResolvedProviderModel, ...] = ()
+
+    def target_at(self, index: int) -> ResolvedProviderModel | None:
+        if index < 0 or index >= len(self.all_targets):
+            return None
+        return self.all_targets[index]
+
+
+@dataclass(frozen=True, slots=True)
+class ResolvedProviderConfig:
+    model: str | None = None
+    provider_fallback: ProviderFallbackConfig | None = None
+    active_target: ResolvedProviderModel = ResolvedProviderModel()
+    target_chain: ResolvedProviderChain = ResolvedProviderChain()
