+++ src/voidcode/provider/resolution.py
@@ -1,0 +1,96 @@
+from __future__ import annotations
+
+from .config import ProviderFallbackConfig
+from .errors import format_invalid_provider_config_error
+from .models import (
+    ProviderModelSelection,
+    ResolvedProviderChain,
+    ResolvedProviderConfig,
+    ResolvedProviderModel,
+)
+from .registry import ModelProviderRegistry
+
+
+def resolve_provider_model(
+    raw_model: str | None,
+    *,
+    registry: ModelProviderRegistry,
+) -> ResolvedProviderModel:
+    if raw_model is None:
+        return ResolvedProviderModel()
+
+    provider_name, model_name = _parse_model_reference(raw_model)
+    provider = registry.resolve(provider_name)
+    return ResolvedProviderModel(
+        selection=ProviderModelSelection(
+            raw_model=raw_model,
+            provider=provider_name,
+            model=model_name,
+        ),
+        provider=provider,
+    )
+
+
+def resolve_provider_chain(
+    provider_fallback: ProviderFallbackConfig | None,
+    *,
+    registry: ModelProviderRegistry,
+) -> ResolvedProviderChain:
+    if provider_fallback is None:
+        return ResolvedProviderChain()
+
+    preferred = resolve_provider_model(provider_fallback.preferred_model, registry=registry)
+    fallbacks = tuple(
+        resolve_provider_model(raw_model, registry=registry)
+        for raw_model in provider_fallback.fallback_models
+    )
+    return ResolvedProviderChain(
+        preferred=preferred,
+        fallbacks=fallbacks,
+        all_targets=(preferred, *fallbacks),
+    )
+
+
+def resolve_provider_config(
+    model: str | None,
+    provider_fallback: ProviderFallbackConfig | None,
+    *,
+    registry: ModelProviderRegistry,
+) -> ResolvedProviderConfig:
+    if provider_fallback is not None:
+        if model is not None and model != provider_fallback.preferred_model:
+            raise ValueError(
+                format_invalid_provider_config_error(
+                    "provider_fallback.preferred_model",
+                    "must match model when both are configured",
+                )
+            )
+        target_chain = resolve_provider_chain(provider_fallback, registry=registry)
+        return ResolvedProviderConfig(
+            model=provider_fallback.preferred_model,
+            provider_fallback=provider_fallback,
+            active_target=target_chain.preferred,
+            target_chain=target_chain,
+        )
+
+    if model is None:
+        return ResolvedProviderConfig()
+
+    active_target = resolve_provider_model(model, registry=registry)
+    target_chain = ResolvedProviderChain(
+        preferred=active_target,
+        all_targets=(active_target,),
+    )
+    return ResolvedProviderConfig(
+        model=model,
+        provider_fallback=None,
+        active_target=active_target,
+        target_chain=target_chain,
+    )
+
+
+def _parse_model_reference(raw_model: str) -> tuple[str, str]:
+    provider_name, separator, model_name = raw_model.partition("/")
+    if separator != "/" or "/" in model_name or not provider_name or not model_name:
+        raise ValueError("model must use provider/model format")
+    return provider_name, model_name
