+++ src/voidcode/runtime/service.py
@@ -12,6 +12,23 @@ from ..graph.contracts import GraphEvent, GraphRunRequest
 from ..graph.read_only_slice import DeterministicReadOnlyGraph
 from ..graph.single_agent_slice import ProviderSingleAgentGraph
 from ..hook.executor import HookExecutionOutcome, HookExecutionRequest, run_tool_hooks
+from ..provider.errors import (
+    SingleAgentContextLimitError,
+    classify_provider_error,
+    format_fallback_exhausted_error,
+    format_invalid_provider_config_error,
+)
+from ..provider.models import (
+    ResolvedProviderChain,
+    ResolvedProviderConfig,
+    ResolvedProviderModel,
+)
+from ..provider.registry import ModelProviderRegistry
+from ..provider.resolution import resolve_provider_config
+from ..provider.snapshot import (
+    parse_resolved_provider_snapshot,
+    resolved_provider_snapshot,
+)
 from ..tools.contracts import Tool, ToolCall, ToolDefinition, ToolResult, ToolResultStatus
 from .acp import AcpAdapter, AcpRequestEnvelope, AcpResponseEnvelope, build_acp_adapter
 from .config import (
@@ -40,27 +57,12 @@ from .events import (
 )
 from .lsp import LspManager, LspManagerState, LspRequest, LspRequestResult, build_lsp_manager
 from .mcp import McpManager, build_mcp_manager
-from .model_provider import (
-    ModelProviderRegistry,
-    ResolvedProviderChain,
-    ResolvedProviderConfig,
-    ResolvedProviderModel,
-    parse_resolved_provider_snapshot,
-    resolve_provider_config,
-    resolved_provider_snapshot,
-)
 from .permission import (
     PendingApproval,
     PermissionDecision,
     PermissionPolicy,
     PermissionResolution,
     resolve_permission,
-)
-from .provider_errors import (
-    SingleAgentContextLimitError,
-    classify_provider_error,
-    format_fallback_exhausted_error,
-    format_invalid_provider_config_error,
 )
 from .session import SessionRef, SessionState, SessionStatus, StoredSessionSummary
 from .single_agent_provider import ProviderExecutionError
